--- packages/components/src/BounceLoader/index.js
@@ -1,1 +1,0 @@
-export { default } from './BounceLoader';
