+++ packages/components/src/LoadingIndicator/LoadingIndicator.story.js
@@ -2,11 +2,11 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { withDocs } from 'storybook-readme';
 
-import BounceLoader from './BounceLoader';
+import LoadingIndicator from './LoadingIndicator';
 import README from './README.md';
 
-storiesOf('Components|BounceLoader', module)
+storiesOf('Components|LoadingIndicator', module)
   .addDecorator(withDocs(README))
   .add('default', () => (
-    <BounceLoader />
+    <LoadingIndicator />
   ));
